Use `Environment.TickCount64` as added here: https://github.com/nanoframework/nanoFramework.Runtime.Native/blob/main/nanoFramework.Runtime.Native/System/Environment.cs

+++ M2Mqtt/Messages/MqttMsgContext.cs
@@ -43,7 +43,7 @@ namespace nanoFramework.M2Mqtt.Messages
         /// <summary>
         /// Timestamp in ticks (for retry)
         /// </summary>
-        public int Timestamp { get; set; }
+        public float Timestamp { get; set; }
 
         /// <summary>
         /// Attempt (for retry)
